requirements.txt
aerich==0.8.0
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
asyncclick==8.1.7.2
asyncio-redis==0.16.0
asyncpg==0.30.0
attrs==24.2.0
certifi==2024.8.30
click==8.1.8
colorama==0.4.6
dictdiffer==0.9.0
fastapi==0.115.6
frozenlist==1.5.0
h11==0.14.0
hiredis==3.1.0
idna==3.10
iso8601==2.1.0
magic-filter==1.0.12
marshmallow==3.23.1
multidict==6.1.0
packaging==24.2
propcache==0.2.1
pydantic==2.9.2
pydantic_core==2.23.4
pypika-tortoise==0.3.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
setuptools==75.7.0
sniffio==1.3.1
starlette==0.41.3
tomlkit==0.13.2
tortoise-orm==0.22.2
typing_extensions==4.12.2
uvicorn==0.34.0
yarl==1.18.3
